Allows to initially collapse the search screenlet of a layered lookup, default to false.
It was harder than I though because of the async nature of Ajax. So I had to use a setTimeout with a 400 ms timeout. Hopefully it will work in any environments. I chose 400 because I found it worked well on my machine, 300 also but to be sure I picked 400, 200 was too low it seems.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/widget@950481 13f79535-47bb-0310-9956-ffa450edef68