Added the <include-portal-page> widget. OFBIZ-3625 https://issues.apache.org/jira/browse/OFBIZ-3625
Using this widget, a PortalPage can now be easily rendered in any screen.
At the bottom of the page an "Edit ON/OFF" button can be used to switch the PortalPage edit mode ON and OFF so that the Portal Page can be edited "in place".
All the portlet editing requests are handled using the "save-last-view" request attribute so that a dedicated screen can be shown to set the Portlet attributes and the column width and, when submitted, the user is taken back to the portal page he was editing.

Those new requests will be used in a later commit to eliminate all the duplicated ****Adm requests that have been introduced when the Webtools portal admin screen has been created.
A simple use of the widget has been added in the Example application.

TODO:
- At the moment the portal page editing is performed on the portal page itself and no portal derivation from the "main" one is performed. This will be done in a later commit.
- The portlet moving controls both manually and by drag and drop have not been implemented yet.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/widget@998570 13f79535-47bb-0310-9956-ffa450edef68