A patch from Leon for "slight patch to "other-field-size" of drop-down form element"" https://issues.apache.org/jira/browse/OFBIZ-5285
"other-field-size" of drop-down form element is a very nice feature, but it's crippled in current version.
attached a patch to make it works and added an example section for that into FormWidgetExample.

jleroux: Interesting, I did not know this feature existed, it was committed long ago by David (before ASF era) 

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/widget@1511279 13f79535-47bb-0310-9956-ffa450edef68