A patch from Youssef Khaye updated by Wei Zhang for "Adding a new attribute ignore-when (verry similar to use-when) that enable rendering or not a field in list or multi form." https://issues.apache.org/jira/browse/OFBIZ-4481
Or on a form of type list or multi, it is not possible to completely ignore a field via use-when attribute. Instead it hides the title and the column header became shifted from the table content.
For example I want to hide the description column from my form when exampleName is choosen as sort-field.

This also modifies
    example/widget/example/ExampleForms.xml#ListExamples
to illustrate this functionality


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/widget@1680865 13f79535-47bb-0310-9956-ffa450edef68