Implements "The lookup and text-area have a read-only attribute but not the text input field" at https://issues.apache.org/jira/browse/OFBIZ-6395
Adds a read-only attribute to the text input field. Using the display type is not enough, because it's uneasy to assign values through Javascript as using the lookup description-field-name shows OOTB. And the disabled option is not good when you want to still show the submitted values on return.

Also 
* improves widget-form.xsd documentation about read-only attribute for lookup
* completes the foFormMacroLibrary.ftl where placeholder tabindex  were missing (if this even ever makes sense)


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/widget@1680921 13f79535-47bb-0310-9956-ffa450edef68