A patch from Christian Carlow for "Form widget <auto-fields-entity default-field-type="find"/> does not render range find fields for fixed-point" https://issues.apache.org/jira/browse/OFBIZ-6443
A form widget with auto-entity-fields for OrderItem will render the quantity fields as text find fields but the prices as range find fields. The quantity find field should match the price range find capability.

This patch adds "fixed-point" to the condition that renders RangeFindField for induceFieldInfoFromEntityField, with support RangeFindField rendering for currency-precise fields also.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/widget@1710177 13f79535-47bb-0310-9956-ffa450edef68