Fixes: Sorting of lists generates undesired results (OFBIZ-8302)
This was due to r1759555 has Scott spotted on. r1759555 fixed a vulnerability 
but as explained in r1759555 commit message we used
>2 redundant mechanisms (better safe than sorry):
>1) linkUrl = URLEncoder.encode(linkUrl, "UTF-8");
>2) sr.append("\" linkUrl=r\"");

Removing the 1st way fixes the reported issue and we are still safe.

I'll have a look at how the catalog/control/FindProduct URL is generated to be 
sure it's OK as is

Thanks: Pierre for report, Scott for spotting the issue.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/widget@1761978 13f79535-47bb-0310-9956-ffa450edef68